Allow protocol write methods to send commands
There exist situations when protocol write_packet/write_data/post_write_interface methods need to write to the interface. Take for example a system that has a command to write to a 64 bit register but setting only the first 8 bits is desired and no mask parameter is provided. The protocol write_packet method may need to send a command to request telemetry containing the current value of the 64 bit register so that the remaining bits of the command can be properly set.